Add Postcode to DTO
When the `MakerPublishedDto` object is deserialized in the
ProductService it requires a `Postcode` value. Before this wasn't
being sent over as I was following the original tutorial.

This means that in the `ProductService`, the database will save
a `Maker` with an `id` and a `postcode` value. It won't know
anything about the `FirstName` or `LastName`.